+++ Cargo.toml
@@ -11,7 +11,7 @@ description = "API for the manga reading tool"
 async-stream = "0.3.6"
 bytes = "1.8.0"
 futures = "0.3.31"
-kdam = "0.5.2"
+kdam = "0.6.0"
 reqwest = { version = "0.12.9", features = ["json"] }
 reqwest-middleware = "0.4.0"
 reqwest-tracing = "0.5.4"
