+++ src/lib.rs
@@ -1,4 +1,5 @@
 // TODO: add bon
+// TODO: remove unnecessary Serialize and Deserialize derives
 
 pub mod requests;
 pub mod viewer;
