+++ src/requests/mod.rs
@@ -3,13 +3,15 @@
 pub mod chapter;
 pub mod cover;
 pub mod manga;
+pub mod query_utils;
 pub mod scanlation_group;
 pub mod tag;
 
-use crate::viewer::{ChapterViewer, DownloadingsSpawnerCommand, ManagerCommand, PageStatus};
+use crate::viewer::PageStatus;
 use chapter::{Chapter, ChapterDownloadMeta};
 use cover::CoverArtAttributes;
 use manga::{Manga, MangaFeedQuery, MangaQuery};
+use query_utils::{EmptyQuery, EntityType, Locale, Query, ResultOk as _};
 use scanlation_group::ScanlationGroup;
 use tag::Tag;
 
@@ -22,313 +24,15 @@ use reqwest::{Client, Response};
 use reqwest_middleware::{ClientBuilder, ClientWithMiddleware};
 use reqwest_tracing::TracingMiddleware;
 
-use tokio::io::AsyncWriteExt as _;
 use tokio::sync::mpsc;
-use tokio::task::{self, JoinSet};
-
+use tokio::task;
 use tokio_stream::wrappers::ReceiverStream;
-use tokio_stream::StreamExt as _;
+use tracing::instrument::Instrument as _;
 
-use tracing::instrument::Instrument;
-
-use std::collections::HashMap;
 use std::default::Default;
 use std::path::PathBuf;
 use std::sync::Mutex;
-use std::sync::{
-    atomic::{AtomicUsize, Ordering},
-    Arc,
-};
-
-pub trait Entity {}
-impl<T: Entity> Entity for Vec<T> {}
-
-pub trait Query: Serialize + std::fmt::Debug {}
-
-#[derive(Serialize, Deserialize, Debug, Clone, Default, Copy)]
-pub struct EmptyQuery {}
-impl Query for EmptyQuery {}
-
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, std::hash::Hash, Clone)]
-#[serde(rename_all = "kebab-case")]
-pub enum Locale {
-    Ab,
-    Aa,
-    Af,
-    Ak,
-    Sq,
-    Am,
-    Ar,
-    An,
-    Hy,
-    Av,
-    Ae,
-    Ay,
-    Az,
-    Bm,
-    Ba,
-    Eu,
-    Be,
-    Bn,
-    Bi,
-    Bs,
-    Br,
-    Bg,
-    My,
-    Ca,
-    Ch,
-    Ce,
-    Ny,
-    Zh,
-    Cu,
-    Cv,
-    Kw,
-    Co,
-    Cr,
-    Hr,
-    Cs,
-    Da,
-    Dv,
-    Dz,
-    En,
-    Eo,
-    Et,
-    Ee,
-    Fo,
-    Fj,
-    Fr,
-    Fi,
-    Fy,
-    Ff,
-    Gd,
-    Gl,
-    Lg,
-    Ka,
-    De,
-    El,
-    Kl,
-    Gn,
-    Gu,
-    Hu,
-    Ht,
-    Ha,
-    He,
-    Hi,
-    Ho,
-    Is,
-    Io,
-    Ig,
-    Id,
-    Ia,
-    Ie,
-    Iu,
-    Ik,
-    Ga,
-    It,
-    Ja,
-    Jv,
-    Kn,
-    Kr,
-    Ks,
-    Kk,
-    Km,
-    Ki,
-    Rw,
-    Ky,
-    Kv,
-    Kg,
-    Ko,
-    Kj,
-    Ku,
-    Lo,
-    La,
-    Lv,
-    Li,
-    Ln,
-    Lt,
-    Lu,
-    Lb,
-    Mk,
-    Mg,
-    Ms,
-    Ml,
-    Mt,
-    Gv,
-    Mi,
-    Mr,
-    Mh,
-    Mn,
-    Na,
-    Nv,
-    Nd,
-    Nr,
-    Ng,
-    Ne,
-    No,
-    Nb,
-    Nn,
-    Nl,
-    Oc,
-    Oj,
-    Or,
-    Om,
-    Os,
-    Pi,
-    Ps,
-    Fa,
-    Pl,
-    Pt,
-    Pa,
-    Qu,
-    Ro,
-    Rm,
-    Rn,
-    Ru,
-    Se,
-    Sm,
-    Sg,
-    Sa,
-    Sc,
-    Sr,
-    Sn,
-    Sd,
-    Si,
-    Sk,
-    Sl,
-    So,
-    St,
-    Es,
-    Su,
-    Sw,
-    Ss,
-    Sv,
-    Tl,
-    Ty,
-    Tg,
-    Ta,
-    Tt,
-    Te,
-    Th,
-    Bo,
-    Ti,
-    To,
-    Ts,
-    Tn,
-    Tr,
-    Tk,
-    Tw,
-    Ug,
-    Uk,
-    Ur,
-    Uz,
-    Ve,
-    Vi,
-    Vo,
-    Wa,
-    Cy,
-    Wo,
-    Xh,
-    Ii,
-    Yi,
-    Yo,
-    Za,
-    Zu,
-    As,
-    ZhHk,
-    PtBr,
-    EsLa,
-    JaRo,
-    KoRo,
-    ZhRo,
-}
-
-pub type LocalizedString = HashMap<Locale, String>;
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(rename_all = "snake_case")]
-pub enum EntityType {
-    Manga,
-    CoverArt,
-    Chapter,
-    Author,
-    ApiClient,
-    ReportReason,
-    ScanlationGroup,
-    User,
-    Tag,
-    Artist,
-    Creator,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(rename_all = "snake_case")]
-pub enum MangaRelation {
-    Monochrome,
-    MainStory,
-    AdaptedFrom,
-    BasedOn,
-    Prequel,
-    SideStory,
-    Doujinshi,
-    SameFranchise,
-    SharedUniverse,
-    Sequel,
-    SpinOff,
-    AlternateStory,
-    AlternateVersion,
-    Preserialization,
-    Colored,
-    Serialization,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct Relationship {
-    id: String,
-    #[serde(rename = "type")]
-    entity_type: EntityType,
-    related: Option<MangaRelation>,
-    attributes: Option<Value>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(rename_all = "camelCase")]
-pub enum Order {
-    Asc,
-    Desc,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, std::hash::Hash)]
-#[serde(rename_all = "camelCase")]
-pub enum OrderOption {
-    Title,
-    Year,
-    CreatedAt,
-    UpdatedAt,
-    LatestUploadedChapter,
-    FollowedCount,
-    Relevance,
-    Chapter,
-}
-
-pub type SortingOptions = HashMap<OrderOption, Order>;
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(rename_all = "snake_case")]
-pub enum ContentRating {
-    Safe,
-    Suggestive,
-    Erotica,
-    Pornographic,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(rename_all = "snake_case")]
-pub enum PublicationDemographic {
-    Shounen,
-    Shoujo,
-    Josei,
-    Seinen,
-}
+use std::sync::{atomic::AtomicUsize, Arc};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct BadResponseError {
@@ -359,29 +63,8 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-// TODO: rename this
-pub trait ResultOk {
-    fn result_ok(&self) -> Result<bool>;
-}
-
-impl ResultOk for Value {
-    fn result_ok(&self) -> Result<bool> {
-        let result = match self.get("result") {
-            Some(status) => status,
-            None => return Err(Error::ParseError),
-        };
-
-        let responded_without_errors = if result.is_string() {
-            let result = result.as_str().expect("verified to be a string");
-
-            result == "ok"
-        } else {
-            return Err(Error::ParseError);
-        };
-
-        Ok(responded_without_errors)
-    }
-}
+pub trait Entity {}
+impl<T: Entity> Entity for Vec<T> {}
 
 #[derive(Clone, Debug)]
 pub struct MangoClient {
@@ -625,387 +308,6 @@ impl MangoClient {
         Ok(resp)
     }
 
-    fn determine_next_download_page(
-        statuses: &mut std::sync::MutexGuard<'_, Vec<PageStatus>>,
-        opened_page: usize,
-    ) -> (Option<usize>, bool) {
-        let mut found_loading_pages = false;
-
-        match statuses[opened_page - 1] {
-            PageStatus::Idle => (Some(opened_page), found_loading_pages),
-            _ => {
-                let mut page = opened_page;
-
-                while page != opened_page - 1 {
-                    if page == statuses.len() {
-                        page = 0;
-
-                        if page == opened_page - 1 {
-                            break;
-                        }
-                    }
-
-                    match statuses[page] {
-                        PageStatus::Idle => {
-                            break;
-                        }
-                        PageStatus::Loading(_) => {
-                            found_loading_pages = true;
-                        }
-                        _ => {}
-                    };
-
-                    page += 1;
-                }
-
-                if page == opened_page - 1 {
-                    if let PageStatus::Loading(_) = statuses[page] {
-                        found_loading_pages = true;
-                    }
-
-                    (None, found_loading_pages)
-                } else {
-                    statuses[page] = PageStatus::Loading(0);
-                    (Some(page + 1), found_loading_pages)
-                }
-            }
-        }
-    }
-
-    #[tracing::instrument(skip(
-        statuses,
-        set_command_sender,
-        downloadings_sender,
-        command_receiver
-    ))]
-    async fn downloadings_manager(
-        opened_page: Arc<AtomicUsize>,
-        statuses: Arc<Mutex<Vec<PageStatus>>>,
-        set_command_sender: mpsc::Sender<DownloadingsSpawnerCommand>,
-        downloadings_sender: Option<mpsc::Sender<usize>>,
-        mut command_receiver: ReceiverStream<ManagerCommand>,
-        max_concurrent_downloads: usize,
-        chapter_size: usize,
-    ) {
-        {
-            let mut statuses = statuses.lock().expect("mutex poisoned");
-
-            for i in 0..chapter_size.min(max_concurrent_downloads) {
-                statuses[i] = PageStatus::Loading(0);
-            }
-        }
-
-        for i in 0..chapter_size.min(max_concurrent_downloads) {
-            set_command_sender
-                .send(DownloadingsSpawnerCommand::NewDownload { page_num: i + 1 })
-                .await
-                .expect("task spawner shutdowned before downloading all pages");
-        }
-
-        while let Some(command) = command_receiver.next().await {
-            tracing::debug!("got command: \n{command:#?}\n");
-            tracing::trace!("\nstatuses: {statuses:#?}\n");
-
-            match command {
-                ManagerCommand::SwitchPage { page_num } => {
-                    opened_page.store(page_num, Ordering::Release);
-                }
-                ManagerCommand::DownloadError { page_num } => {
-                    let new_download_command = DownloadingsSpawnerCommand::NewDownload { page_num };
-                    set_command_sender
-                        .send(new_download_command)
-                        .await
-                        .expect("join_set task shutdowned before downloading all pages");
-
-                    tracing::trace!("manager sent {new_download_command:#?} command to set");
-                }
-                ManagerCommand::DownloadedSuccessfully { page_num } => {
-                    // NOTE: we don't care if the receiver is no longer interested in this
-                    // information
-                    if let Some(sender) = &downloadings_sender {
-                        let _ = sender.send(page_num).await;
-                    };
-
-                    // TODO: i'm not sure if this scope is required
-                    let (next_download_page, found_loading_pages) = {
-                        Self::determine_next_download_page(
-                            &mut statuses.lock().expect("mutex poisoined"),
-                            opened_page.load(Ordering::Acquire),
-                        )
-                    };
-
-                    match next_download_page {
-                        Some(page) => {
-                            let command =
-                                DownloadingsSpawnerCommand::NewDownload { page_num: page };
-
-                            set_command_sender
-                                .send(command)
-                                .await
-                                .expect("join_set task shutdowned before downloading all pages");
-
-                            tracing::trace!("manager sent {command:#?} command to set");
-                        }
-                        None => {
-                            if !found_loading_pages {
-                                set_command_sender
-                                    .send(DownloadingsSpawnerCommand::Shutdown)
-                                    .await
-                                    .expect(
-                                        "join_set task shutdowned before the respectful command",
-                                    );
-
-                                tracing::trace!("manager sent shutdown command to set");
-
-                                break;
-                            }
-                        }
-                    };
-                }
-            }
-        }
-        tracing::debug!("shutdowned");
-    }
-
-    async fn handle_downloading_page_request_error(
-        e: reqwest::Error,
-        manager_command_sender: &mpsc::Sender<ManagerCommand>,
-        page_num: usize,
-    ) {
-        // TODO: handle possible errors
-
-        tracing::warn!("got respond from the server: {e:#?}");
-
-        manager_command_sender
-            .send(ManagerCommand::DownloadError { page_num })
-            .await
-            .expect("manager shutdowned before getting the signal from last downloading task");
-    }
-
-    /// Queries next chunk, handles possible errors and returnes this chunk if no errors
-    /// encountered or None if errors were met or the source is drained along with the boolean
-    /// value that represents if errores were emitted
-    async fn get_next_chunk(
-        resp: &mut Response,
-        manager_command_sender: &mpsc::Sender<ManagerCommand>,
-        page_num: usize,
-    ) -> (Option<Bytes>, bool) {
-        match resp.chunk().in_current_span().await {
-            Ok(chunk) => {
-                let with_errors = false;
-                (chunk, with_errors)
-            }
-            Err(e) => {
-                Self::handle_downloading_page_request_error(e, manager_command_sender, page_num)
-                    .in_current_span()
-                    .await;
-
-                let with_errors = true;
-                (None, with_errors)
-            }
-        }
-    }
-
-    #[tracing::instrument(skip(client, statuses, manager_command_sender))]
-    async fn downloading_page(
-        client: MangoClient,
-        chapter_hash: String,
-        page_num: usize,
-        url: String,
-        statuses: Arc<Mutex<Vec<PageStatus>>>,
-        manager_command_sender: mpsc::Sender<ManagerCommand>,
-    ) {
-        let out_page_filename = format!("tmp/{}/{}.png", chapter_hash, page_num);
-
-        let mut out_page = tokio::fs::File::create(&out_page_filename)
-            .await
-            .expect("failed to open file to save page");
-
-        let resp_res = client.get_page_chunks(&url).in_current_span().await;
-
-        match resp_res {
-            Ok(mut resp) => {
-                let total_size = resp
-                    .content_length()
-                    .expect("could not get the content_length of page");
-
-                while let (Some(chunk), with_error) =
-                    Self::get_next_chunk(&mut resp, &manager_command_sender, page_num)
-                        .in_current_span()
-                        .await
-                {
-                    if with_error {
-                        return;
-                    }
-
-                    let cur_size = chunk.len();
-
-                    out_page
-                        .write_all(chunk.as_ref())
-                        .await
-                        .expect("failed to save page");
-
-                    {
-                        let mut statuses = statuses.lock().expect("mutex poisoned");
-                        let already_loaded = match statuses[page_num - 1] {
-                            PageStatus::Loading(percent) => percent,
-                            _ => unreachable!(),
-                        };
-
-                        statuses[page_num - 1] = PageStatus::Loading(
-                            already_loaded + 100 * cur_size / total_size as usize,
-                        );
-                    }
-                }
-
-                {
-                    let mut statuses = statuses.lock().expect("mutex poisoned");
-                    statuses[page_num - 1] = PageStatus::Loaded(out_page_filename.into());
-                }
-
-                manager_command_sender
-                    .send(ManagerCommand::DownloadedSuccessfully { page_num })
-                    .await
-                    .expect(
-                        "manager shutdowned before getting the signal from last downloading task",
-                    );
-            }
-
-            Err(e) => match e {
-                Error::RequestError(e) => {
-                    Self::handle_downloading_page_request_error(
-                        e,
-                        &manager_command_sender,
-                        page_num,
-                    )
-                    .in_current_span()
-                    .await;
-                }
-                _ => {
-                    panic!("Got an unexpected error from download page process")
-                }
-            },
-        }
-    }
-
-    #[tracing::instrument(skip_all)]
-    async fn downloadings_spawner(
-        client: MangoClient,
-        meta: ChapterDownloadMeta,
-        mut command_receiver: ReceiverStream<DownloadingsSpawnerCommand>,
-        manager_command_sender: mpsc::Sender<ManagerCommand>,
-        statuses: Arc<Mutex<Vec<PageStatus>>>,
-    ) {
-        let mut set = JoinSet::new();
-
-        while let Some(command) = command_receiver.next().await {
-            tracing::debug!("got command: \n{command:#?}\n");
-
-            match command {
-                DownloadingsSpawnerCommand::Shutdown => {
-                    set.shutdown().await;
-
-                    break;
-                }
-                DownloadingsSpawnerCommand::NewDownload { page_num } => {
-                    // NOTE: we don't care about the result of the computations, this is simply
-                    // for the sake of buffer not overflowing
-                    let _ = set.try_join_next();
-
-                    let download_url = format!(
-                        "{}/data/{}/{}",
-                        &meta.base_url,
-                        &meta.chapter.hash,
-                        &meta.chapter.data[page_num - 1]
-                    );
-
-                    set.spawn(Self::downloading_page(
-                        client.clone(),
-                        meta.chapter.hash.clone(),
-                        page_num,
-                        download_url,
-                        Arc::clone(&statuses),
-                        manager_command_sender.clone(),
-                    ));
-
-                    tracing::trace!("spawned new download task");
-                }
-            }
-        }
-
-        tracing::debug!("shutdowned");
-    }
-
-    #[tracing::instrument(skip(self))]
-    pub async fn chapter_viewer(
-        &self,
-        chapter_id: &str,
-        mut max_concurrent_downloads: usize,
-    ) -> Result<ChapterViewer> {
-        max_concurrent_downloads = max_concurrent_downloads.max(1);
-
-        let download_meta = self
-            .get_chapter_download_meta(chapter_id)
-            .in_current_span()
-            .await?;
-
-        let chapter_size = download_meta.chapter.data.len();
-        let buf = Arc::new(Mutex::new(vec![PageStatus::Idle; chapter_size]));
-
-        let (downloadings_sender, downloading_receiver) = mpsc::channel(chapter_size);
-        let downloading_receiver = ReceiverStream::new(downloading_receiver);
-
-        let (manager_command_sender, manager_command_receiver) = mpsc::channel(10);
-        let manager_command_receiver = ReceiverStream::new(manager_command_receiver);
-
-        let (downloadings_spawner_command_sender, downloadings_spawner_command_receiver) =
-            mpsc::channel(10);
-        let downloadings_spawner_command_receiver =
-            ReceiverStream::new(downloadings_spawner_command_receiver);
-
-        let res = ChapterViewer {
-            opened_page: Arc::new(AtomicUsize::new(1)),
-            downloadings: downloading_receiver,
-            meta: download_meta,
-            statuses: buf,
-            submit_switch: manager_command_sender.clone(),
-        };
-
-        let chapter_download_dir = format!("tmp/{}", res.meta.chapter.hash);
-        if !std::fs::exists(&chapter_download_dir).expect("failed to get info about tmp directory")
-        {
-            std::fs::create_dir_all(&chapter_download_dir)
-                .expect("failed to create directory for storing pages");
-        } else {
-            let dir_meta = std::fs::metadata(&chapter_download_dir)
-                .expect("failed to create directory for storing pages");
-
-            if !dir_meta.is_dir() {
-                panic!("failed to create directory for storing pages: file already exists, not a directory");
-            }
-        }
-
-        task::spawn(Self::downloadings_manager(
-            Arc::clone(&res.opened_page),
-            Arc::clone(&res.statuses),
-            downloadings_spawner_command_sender.clone(),
-            Some(downloadings_sender),
-            manager_command_receiver,
-            max_concurrent_downloads,
-            chapter_size,
-        ));
-
-        task::spawn(Self::downloadings_spawner(
-            self.clone(),
-            res.meta.clone(),
-            downloadings_spawner_command_receiver,
-            manager_command_sender,
-            Arc::clone(&res.statuses),
-        ));
-
-        return Ok(res);
-    }
-
     #[tracing::instrument(skip(self))]
     pub async fn download_full_chapter(
         &self,
@@ -1081,6 +383,11 @@ mod tests {
     use tracing_subscriber::filter::LevelFilter;
     use tracing_subscriber::layer::SubscriberExt;
     use tracing_subscriber::prelude::*;
+
+    use query_utils::{Order, OrderOption};
+    use std::collections::HashMap;
+
+    use tokio::io::AsyncWriteExt as _;
 
     #[tokio::test]
     async fn test_search_manga() {
@@ -1374,7 +681,7 @@ mod tests {
 
         let mut page_paths = Vec::new();
         for i in 0..chapter_len {
-            page_paths.push(viewer.get_page(i + 1).await);
+            page_paths.push(viewer.open_page(i + 1).await);
         }
 
         let mut out = tokio::fs::File::create("test_files/page_paths")
